Game of Tic Tac Toe
You can make the board as big as as you want and also the winning sequence can be increased. That is in the traditional game of Tic Tac Toe the winning sequence in 3 X's in a line or 3 0's in a line. In this game that number can be changed . Also in traditional tic tac toe only 2 players can play. Here a maximum of 26 players can play!

You can save a game by pressing S for save and even load a saved game.
Press Q to Quit